# docs/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Little Bear RAG Backend API
  version: 0.1.0
  description: >
    P0 OpenAPI contract for Little Bear RAG backend. The contract covers setup,
    authentication, organization, permission, knowledge base, document import,
    indexing operations, query, audit, healthcheck, config, and internal model
    gateway APIs. P0 uses external embedding, rerank, and LLM provider
    endpoints instead of built-in model simulators, and does not include complete
    config approval or rollback workflows.

    Backend contract paths use /internal/v1. OpenAPI, backend routes, frontend
    SDKs, and contract tests should all use this path prefix consistently.
servers:
  - url: http://localhost:8000
    description: Local development
...
security:
  - bearerAuth: []
paths:
...
  /internal/v1/admin/audit-logs:
    get:
      tags: [AdminAudit]
      summary: Admin list audit logs
      operationId: adminListAuditLogs
      x-token-type: access
      x-required-scope: audit:read
      x-audit-required: true
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
        - name: actor_id
          in: query
          schema:
            type: string
        - name: action
          in: query
          schema:
            type: string
        - name: resource_type
          in: query
          schema:
            type: string
        - name: result
          in: query
          schema:
            type: string
        - name: risk_level
          in: query
          schema:
            $ref: "#/components/schemas/RiskLevel"
      responses:
        "200":
          description: Audit logs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuditLogListResponse"
        default:
          $ref: "#/components/responses/Error"
  /internal/v1/admin/audit-logs/{audit_id}:
    get:
      tags: [AdminAudit]
      summary: Admin get audit log
      operationId: adminGetAuditLog
      x-token-type: access
      x-required-scope: audit:read
      x-audit-required: true
      parameters:
        - $ref: "#/components/parameters/AuditId"
      responses:
        "200":
          description: Audit log
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuditLogResponse"
        default:
          $ref: "#/components/responses/Error"
  /internal/v1/admin/query-logs:
    get:
      tags: [AdminAudit]
      summary: Admin list query logs
      operationId: adminListQueryLogs
      x-token-type: access
      x-required-scope: audit:read
      x-audit-required: true
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
        - name: user_id
          in: query
          schema:
            type: string
        - name: kb_id
          in: query
          schema:
            type: string
        - name: degraded
          in: query
          schema:
            type: boolean
        - name: status
          in: query
          schema:
            type: string
        - name: degrade_reason
          in: query
          schema:
            type: string
        - name: request_id
          in: query
          schema:
            type: string
        - name: trace_id
          in: query
          schema:
            type: string
        - name: error_code
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Query logs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QueryLogListResponse"
        default:
          $ref: "#/components/responses/Error"
  /internal/v1/admin/query-logs/{query_log_id}:
    get:
      tags: [AdminAudit]
      summary: Admin get query log
      operationId: adminGetQueryLog
      x-token-type: access
      x-required-scope: audit:read
      x-audit-required: true
      parameters:
        - $ref: "#/components/parameters/QueryLogId"
      responses:
        "200":
          description: Query log
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QueryLogResponse"
        default:
          $ref: "#/components/responses/Error"
  /internal/v1/admin/model-call-logs:
    get:
      tags: [AdminAudit]
      summary: Admin list model call logs
      operationId: adminListModelCallLogs
      x-token-type: access
      x-required-scope: audit:read
      x-audit-required: true
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
        - name: model
          in: query
          schema:
            type: string
        - name: degraded
          in: query
          schema:
            type: boolean
        - name: model_type
          in: query
          schema:
            type: string
        - name: caller
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
        - name: request_id
          in: query
          schema:
            type: string
        - name: trace_id
          in: query
          schema:
            type: string
        - name: error_code
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Model call logs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ModelCallLogListResponse"
        default:
          $ref: "#/components/responses/Error"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSize:
      name: page_size
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 20
...
    AuditId:
      name: audit_id
      in: path
      required: true
      schema:
        type: string
    QueryLogId:
      name: query_log_id
      in: path
      required: true
      schema:
        type: string
  responses:
    Error:
      description: Error response
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    RequestId:
      type: string
      examples: ["req_001"]
...
    Timestamp:
      type: string
      format: date-time
    AnyObject:
      type: object
      additionalProperties: true
    Pagination:
      type: object
      required: [page, page_size, total]
      properties:
        page:
          type: integer
        page_size:
          type: integer
        total:
          type: integer
    ErrorResponse:
      type: object
      required: [request_id, error_code, message, retryable]
      properties:
        request_id:
          $ref: "#/components/schemas/RequestId"
        error_code:
          type: string
        message:
          type: string
        stage:
          type: string
        retryable:
          type: boolean
        details:
          $ref: "#/components/schemas/AnyObject"
...
    ResourceType:
      type: string
      enum: [enterprise, department, user, role, role_binding, permission, knowledge_base, folder, document, chunk, import_job, config, query, setup, model_call]
    RiskLevel:
      type: string
      enum: [low, medium, high, critical]
...
    AuditLog:
      type: object
      properties:
        id:
          type: string
        request_id:
          type: [string, "null"]
        trace_id:
          type: [string, "null"]
        event_name:
          type: string
        actor_type:
          type: string
        actor_id:
          type: [string, "null"]
        action:
          type: string
        resource_type:
          $ref: "#/components/schemas/ResourceType"
        resource_id:
          type: [string, "null"]
        result:
          type: string
        risk_level:
          $ref: "#/components/schemas/RiskLevel"
        config_version:
          type: [integer, "null"]
        permission_version:
          type: [integer, "null"]
        index_version_hash:
          type: [string, "null"]
        summary_json:
          type: object
          additionalProperties: true
        error_code:
          type: [string, "null"]
        created_at:
          $ref: "#/components/schemas/Timestamp"
    AuditLogResponse:
      type: object
      properties:
        request_id:
          $ref: "#/components/schemas/RequestId"
        data:
          $ref: "#/components/schemas/AuditLog"
    AuditLogListItem:
      type: object
      properties:
        id:
          type: string
        event_name:
          type: string
        actor_type:
          type: string
        action:
          type: string
        resource_type:
          $ref: "#/components/schemas/ResourceType"
        result:
          type: string
        risk_level:
          $ref: "#/components/schemas/RiskLevel"
        config_version:
          type: [integer, "null"]
        permission_version:
          type: [integer, "null"]
        error_code:
          type: [string, "null"]
        created_at:
          $ref: "#/components/schemas/Timestamp"
    AuditLogListResponse:
      type: object
      properties:
        request_id:
          $ref: "#/components/schemas/RequestId"
        data:
          type: array
          items:
            $ref: "#/components/schemas/AuditLogListItem"
        pagination:
          $ref: "#/components/schemas/Pagination"
    QueryLog:
      type: object
      properties:
        id:
          type: string
        request_id:
          type: string
        trace_id:
          type: string
        user_id:
          type: string
        kb_ids:
          type: array
          items:
            type: string
        query_hash:
          type: string
        status:
          type: string
        degraded:
          type: boolean
        degrade_reason:
          type: [string, "null"]
        config_version:
          type: integer
        permission_version:
          type: integer
        permission_filter_hash:
          type: string
        index_version_hash:
          type: [string, "null"]
        model_route_hash:
          type: [string, "null"]
        latency_ms:
          type: integer
        candidate_count:
          type: integer
        citation_count:
          type: integer
        error_code:
          type: [string, "null"]
        created_at:
          $ref: "#/components/schemas/Timestamp"
    QueryLogResponse:
      type: object
      properties:
        request_id:
          $ref: "#/components/schemas/RequestId"
        data:
          $ref: "#/components/schemas/QueryLog"
    QueryLogListResponse:
      type: object
      properties:
        request_id:
          $ref: "#/components/schemas/RequestId"
        data:
          type: array
          items:
            $ref: "#/components/schemas/QueryLog"
        pagination:
          $ref: "#/components/schemas/Pagination"
    ModelCallLog:
      type: object
      properties:
        id:
          type: string
        request_id:
          type: [string, "null"]
        trace_id:
          type: string
        caller:
          type: string
        model_type:
          type: string
        model_name:
          type: string
        model_version:
          type: [string, "null"]
        model_route_hash:
          type: string
        status:
          type: string
        latency_ms:
          type: integer
        token_usage_json:
          type: [object, "null"]
          additionalProperties: true
        degraded:
          type: boolean
        config_version:
          type: [integer, "null"]
        prompt_hash:
          type: [string, "null"]
        input_hash:
          type: [string, "null"]
        output_hash:
          type: [string, "null"]
        error_code:
          type: [string, "null"]
        created_at:
          $ref: "#/components/schemas/Timestamp"
    ModelCallLogListResponse:
      type: object
      properties:
        request_id:
          $ref: "#/components/schemas/RequestId"
        data:
          type: array
          items:
            $ref: "#/components/schemas/ModelCallLog"
        pagination:
          $ref: "#/components/schemas/Pagination"
...